Move compile command into its own file
