+++ lammps_testing/lammps_test.py
@@ -7,15 +7,14 @@ import subprocess
 import sys
 
 from termcolor import colored
-from .common import logger, Settings, Container, discover_tests, get_containers, get_configurations, get_container, get_configuration, get_names
-from .tests import CompilationTest, RunTest, RegressionTest, UnitTest
+from .common import logger, Settings, Container, get_containers, get_configurations, get_container, get_configuration
+from .tests import RunTest, RegressionTest, UnitTest
 
 from .commands.env import init_command as init_env_command
 from .commands.config import init_command as init_config_command
 from .commands.build import init_command as init_build_command
+from .commands.compile import init_command as init_compile_command
 from .commands.reg import init_command as init_reg_command
-
-
 
 class TestRunner():
     def __init__(self, builder, settings):
@@ -189,69 +188,6 @@ def checkstyle(args, settings):
             print(f"{filename}:exception while reading file:{e}")
 
 
-
-def expand_selected_config_and_builds(selected_builds, settings):
-    builds = {}
-
-    def add_build(config, build):
-        if config not in builds:
-            builds[config] = [build]
-        elif build not in builds[config]:
-            builds[config].append(build)
-
-    if "ALL" in selected_builds or "all" in selected_builds:
-        configurations = get_configurations(settings)
-        for config in configurations:
-            if hasattr(config, "builds"):
-                builds[config.name] = config.builds
-    else:
-        for selected in selected_builds:
-            parts = selected.split('/')
-
-            if len(parts) == 1 or parts[1] == "*":
-                config = get_configuration(parts[0], settings)
-                for build in config.builds:
-                    add_build(config.name, build)
-            elif parts[1].endswith("*"):
-                prefix = parts[1][:-1]
-                config = get_configuration(parts[0], settings)
-                for build in config.builds:
-                    if build.startswith(prefix):
-                        add_build(config.name, build)
-            else:
-                config = get_configuration(parts[0], settings)
-                build = parts[1]
-                add_build(config.name, build)
-    return builds
-
-
-
-def compilation_test(args, settings):
-    selected_builds = args.builds
-    selected_builds = expand_selected_config_and_builds(args.builds, settings)
-    configurations = get_configurations(settings)
-
-    for config in configurations:
-        if config.name in selected_builds.keys():
-            container = get_container(config.container_image, settings)
-
-            if not container.exists:
-                logger.error(f"Can not find container: {container}")
-                sys.exit(1)
-
-            for build_name in config.builds:
-                if build_name in selected_builds[config.name]:
-                    if args.ignore_commit:
-                        compTest = CompilationTest(build_name, container, settings)
-                    else:
-                        compTest = CompilationTest(build_name, container, settings, settings.current_lammps_commit)
-
-                    print(f"Compiling {config.name}/{build_name}")
-
-                    if not compTest.build():
-                        print(f"Compilation of '{build_name}' on '{container.name}' FAILED!")
-                        sys.exit(1)
-
 def regression_test(args, settings):
     selected_config = args.config
     selected_builds = args.builds
@@ -362,6 +298,10 @@ def main():
     parser_build = subparsers.add_parser('build', help='test build commands')
     init_build_command(parser_build)
 
+    # create the parser for the "compile" command
+    parser_compile = subparsers.add_parser('compile', help='run compilation tests')
+    init_compile_command(parser_compile)
+
     # create the parser for the "reg" command
     parser_build = subparsers.add_parser('reg', help='test reg commands')
     init_reg_command(parser_build)
@@ -374,12 +314,6 @@ def main():
     # create the parser for the "checkstyle" command
     parser_checkstyle = subparsers.add_parser('checkstyle', help='check current checkout for code style issues')
     parser_checkstyle.set_defaults(func=checkstyle)
-
-    # create the parser for the "compile" command
-    parser_compile_test = subparsers.add_parser('compile', help='run compilation tests')
-    parser_compile_test.add_argument('--ignore-commit', default=False, action='store_true', help='Ignore commit and do not create SHA specific build folder')
-    parser_compile_test.add_argument('builds', metavar='build', nargs='*', default=['ALL'], help='list of builds that should be compiled')
-    parser_compile_test.set_defaults(func=compilation_test)
 
     # create the parser for the "run" command
     parser_run_test = subparsers.add_parser('run', help='run tests')
